Move PlatformLayers samples to Dialogs page. But temporary saves references in PlatformLayers page as well.
