+++ compose/mpp/demo/src/commonMain/kotlin/androidx/compose/mpp/demo/layers/PlatformLayers.kt
@@ -1,0 +1,31 @@
+/*
+ * Copyright 2024 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package androidx.compose.mpp.demo.layers
+
+import androidx.compose.mpp.demo.Screen
+import androidx.compose.mpp.demo.components.dialog.DialogWithTextField
+import androidx.compose.mpp.demo.components.dialog.FocusAndKeyInput
+
+/**
+ * Test cases for platform layers
+ * TODO remove after stabilizing platformLayers feature
+ */
+val PlatformLayers = Screen.Selection(
+    "PlatformLayers",
+    DialogWithTextField,
+    FocusAndKeyInput,
+)
